Add ajp10304 community layout.
