+++ layouts/community/ortho_4x12/ajp10304/rules.mk
@@ -1,0 +1,3 @@
+AUDIO_ENABLE = no
+MOUSEKEY_ENABLE = yes
+DYNAMIC_MACRO_ENABLE = yes
