+++ users/ajp10304/rules.mk
@@ -1,5 +1,3 @@
 SRC += ajp10304.c
 
-AUDIO_ENABLE = no
-MOUSEKEY_ENABLE = yes
-DYNAMIC_MACRO_ENABLE = yes
+
